requirements.txt
fastapi
uvicorn
httpx
python-dotenv
ipython
soundfile